requirements.txt
Django==4.2
openpyxl==3.1.5
django-user-agents==0.4.0
python-dotenv==1.0.1